fix relative bugs in extractAllFiles

+++ Data Preprocess/src/extractAllFiles.py
@@ -9,17 +9,20 @@ def deleteEmptyHintsFile(path, b):
         fileName = file[:file.find('.hints')]
         hintPath = str(fileName + '.hints')
         parentDirectory=os.path.abspath(os.path.dirname(os.getcwd()))
-        print(fileName)
+        print("fileName:", file)
         #if the hint file is empty delete hint file and other files
         f=open(file,"r")
         content=f.read()
         if not content:
 
-            print("delete",fileName,'.hints/horn.negativeHInts/redundantHints')
+            print("delete",fileName+'.hints/horn.negativeHInts/redundantHints')
             os.remove(fileName+'.hints')
-            os.remove(fileName + '.horn')
-            os.remove(fileName + '.negativeHints')
-            os.remove(fileName + '.redundantHints')
+            if os.path.isfile(fileName + '.horn'):
+                os.remove(fileName + '.horn')
+            if os.path.isfile(fileName + '.negativeHints'):
+                os.remove(fileName + '.negativeHints')
+            if os.path.isfile(fileName + '.redundantHints'):
+                os.remove(fileName + '.redundantHints')
 
         #if there is no hint file delete othr relevant files
         # exists = os.path.isfile(hintPath)
@@ -28,8 +31,8 @@ def deleteEmptyHintsFile(path, b):
         #     os.remove(fileName + '.horn')
         #     os.remove(fileName + '.negativeHints')
         #     os.remove(fileName + '.redundantHints')
-
-
+        #
+        #
 
 
     #os.remove("ChangedFile.csv")
@@ -42,10 +45,10 @@ def main():
     abstractionOption = 'abstract:manual'
 
     benchmarkList = list()
-    benchmarkList.append('svcomp16/locks')
+    #benchmarkList.append('svcomp16/locks')
     #benchmarkList.append('svcomp16/loop-acceleration')
     #benchmarkList.append('svcomp16/loop-invgen')
-    #benchmarkList.append('svcomp16/loop-lit')
+    benchmarkList.append('svcomp16/loop-lit')
     #benchmarkList.append('svcomp16/loop-new')
     #benchmarkList.append('svcomp16/loops')
     #benchmarkList.append('svcomp16/ntdrivers-simplified')
@@ -58,16 +61,18 @@ def main():
     for b in benchmarkList:
         extractHornClausesFromMultipleProgram(filePath, b, abstractionOption)
         extractRedundantHintsFromMultipleProgram(filePath, b, abstractionOption)
+        print()
 
     print('------------------------')
     for b in benchmarkList:
         curpath = os.path.abspath(os.curdir)
-        path = curpath + '/' + b + '/'
+        parenDir = os.path.abspath(os.path.pardir)
+        path = parenDir + '/' + b + '/'
         print(path)
         extractNegativeTrainData(path, b)
         deleteEmptyHintsFile(path, b)
 
-
+    print("Finished")
 
 if __name__ == '__main__':
     main()
